Refactor: characterId 타입 선언 위치를 module로 이동 (#191)
