+++ src/features/characters/containers/CharactersContainer.tsx
@@ -1,7 +1,7 @@
 import { useCreateThread } from '@/src/apis/_generated/serverAPI';
 import { COLOR } from '@/src/constants/theme';
 import { LargeTitle } from '@/src/core/Txt';
-
+import type { CharacterId } from '@/src/modules/character';
 import { toast } from '@/src/modules/toast';
 import { useRouter } from 'expo-router';
 import { useState } from 'react';
@@ -12,7 +12,6 @@ import ActionButtons from '../components/ActionButtons';
 import CharacterList from '../components/CharacterList';
 import CharacterPageHeader from '../components/CharacterPageHeader';
 import { useCharactersQueryParams } from '../hooks/useCharactersQueryParams';
-import type { CharacterId } from '../types';
 
 const CharactersContainer = () => {
   const { year, month, day } = useCharactersQueryParams();
