+++ src/modules/character/types/index.ts
